backend/requirements.txt
fastapi
python-dotenv
SQLAlchemy==1.4.31
pydantic
psycopg2-binary